Basic assimp importer implementation

+++ src/Wasabi/Images/WRenderTarget.cpp
@@ -462,7 +462,7 @@ VkCommandBuffer WRenderTarget::GetCommnadBuffer() const {
 }
 
 uint32_t WRenderTarget::GetNumColorOutputs() const {
-	return !Valid() ? 0 : (m_targets.size() == 0 ? 1 : m_targets.size());
+	return !Valid() ? 0 : (m_targets.size() == 0 ? 1 : (uint32_t)m_targets.size());
 }
 
 bool WRenderTarget::HasDepthOutput() const {
